Pids added to controller class fixed error with sensor fuctions made motor accept doubles instead of ints

+++ Quadrotor/src/Motor.cpp
@@ -25,7 +25,7 @@ Motor::Motor(int motorNumber) {
 //	open(filename, O_RDWR);
 }
 
-void Motor::setPower(int newPower){
+void Motor::setPower(double newPower){
 	power=newPower;
 }
